Create Freestyle project, add description
